Move migrations out of lib/
cc #135
