+++ lib/index.js
@@ -1,4 +1,4 @@
-const models = require('./db/models');
+const models = require('./models');
 const setupSlack = require('./slack');
 const setupGitHub = require('./github');
