+++ lib/models/index.js
@@ -4,7 +4,7 @@ const Sequelize = require('sequelize');
 
 const basename = path.basename(module.filename);
 const env = process.env.NODE_ENV || 'development';
-const config = require('../config.json')[env];
+const config = require('../../db/config.json')[env];
 
 const db = {};
