+++ lib/models/installation.js
