+++ lib/models/slackuser.js
